refactor: remove neuromod module, inline NeuromodNeuron in router
- Move `NeuromodNeuron` into `src/router.rs` as a router-internal NIF primitive
- Delete `src/neuromod.rs` and remove `pub mod neuromod` from `lib.rs`
- Re-export `NeuromodNeuron` from `router` to preserve public API
- Document the `synaptic-mesh`/`neuromod` crate boundary in `lib.rs` and `README.md`
- Update `REVIEW.md` regression-guard paths and add `.devin/blueprint.yaml`
