+++ .devin/blueprint.yaml
@@ -1,0 +1,29 @@
+# Devin environment blueprint for synaptic-mesh.
+#
+# This file is a version-controlled REFERENCE ONLY. Devin blueprints are not
+# auto-read from the repo — copy this content into Devin's
+# Settings > Environment > Blueprints > synaptic-mesh UI to apply it.
+# Keep this file in sync whenever the blueprint is changed in the UI.
+
+initialize: |
+  rustup toolchain install stable --component clippy --component rustfmt
+  rustup default stable
+
+maintenance: |
+  cargo fetch --locked
+
+knowledge:
+  - name: format-check
+    contents: cargo fmt --check
+  - name: test
+    contents: cargo test --locked
+  - name: lint
+    contents: cargo clippy --all-features -- -D warnings
+  - name: review-gate
+    contents: |
+      Before opening or updating a PR, run every command in REVIEW.md
+      (format check, locked test matrix, clippy, regression guards,
+      diff hygiene, origin hygiene). All must pass; REVIEW.md's
+      "Pass criteria" section is the merge bar.
+  - name: repo-map
+    contents: See AGENTS.md for module map, entry order, and design principles.
